Update: Modularise Avatar component
