+++ src/components/alexandria/Avatar/styles.scss
@@ -1,7 +1,7 @@
-@import '../border';
-@import '../color';
-@import '../font-size';
-@import '../font-weight';
+@import 'styles/border';
+@import 'styles/color';
+@import 'styles/font-size';
+@import 'styles/font-weight';
 
 $diameter: 35px;
